requirements.txt
bcrypt
blinker
click
docutils
Flask
iso8601
jsonschema
PyJWT
pytest
requests
rethinkdb
# Shapely==1.5.12
simplegeneric
six
Sphinx==1.2.3
sphinxcontrib-napoleon==0.3.11
Werkzeug
python-slugify